remove test with dummy get-result
Most times, when select resturn to indicate there is a result to fetch
from a async search, but when reading it returns no result. (bonsai
bug?)
This was confirmed by just an extra dummy read. This commit removes
it again, because about 1 in 25 times it does read away the result
and the subsequent 'real' get-result raises an exception for a wrong
msgid.